Finally got fed up with the clunky fluent interface for building XML documents using the library. Now using underscore templates (already have underscore in the app) instead. Put the templates in /private and pull them up as assets, compile, apply template to JSON for interpolation. Don't care about precompiling templates as there is no load.
Start displaying the different API forms in the order needed if the previous
step has been successfully completed based on what data is in the meteor
session. We do not care about resetting everything if you decide to start
over at a much earlier step-- this is POC so click the "reset" buttons from
bottom to top to reset things in the right order.

Too big of a commit.